@@ -1,38 +1,38 @@
 ext {
     projectVersion = [
             minSdk    : 19,
-            targetSdk : 30,
-            compileSdk: 30,
-            kotlin    : "1.4.31",
+            targetSdk : 31,
+            compileSdk: 31,
+            kotlin    : "1.5.20",
     ]
     projectDependency = [
 
             // Gradle Plugins
             kotlinPlugin       : "org.jetbrains.kotlin:kotlin-gradle-plugin:${projectVersion.kotlin}",
-            androidGradlePlugin: "com.android.tools.build:gradle:4.1.2",
+            androidGradlePlugin: "com.android.tools.build:gradle:7.2.0-alpha06",
 
             // Dependencies
             kotlinStdlibJdk8   : "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${projectVersion.kotlin}",
-            appCompat          : "androidx.appcompat:appcompat:1.2.0",
+            kotlinReflect      : "org.jetbrains.kotlin:kotlin-reflect:${projectVersion.kotlin}",
+            appCompat          : "androidx.appcompat:appcompat:1.4.0",
 
             // Test dependencies
             junit              : "junit:junit:4.13.2",
-            truth              : "com.google.truth:truth:1.1.2",
-            supportTestRunner  : "androidx.test:runner:1.2.0",
-            espressoCore       : "androidx.test.espresso:espresso-core:3.3.0",
-            androidJUnit       : "androidx.test.ext:junit:1.1.2",
+            truth              : "com.google.truth:truth:1.1.3",
+            supportTestRunner  : "androidx.test:runner:1.4.0",
+            espressoCore       : "androidx.test.espresso:espresso-core:3.4.0",
+            androidJUnit       : "androidx.test.ext:junit:1.1.3",
             commonsCsv         : "org.apache.commons:commons-csv:1.8",
             kotlinTest         : "org.jetbrains.kotlin:kotlin-test:${projectVersion.kotlin}",
-            robolectric        : "org.robolectric:robolectric:4.5.1",
+            robolectric        : "org.robolectric:robolectric:4.7.3",
     ]
 
     // Used by the plugin-version-handler.gradle
     pluginVersions = [
-            "org.jetbrains.kotlin.android"      : "${projectVersion.kotlin}",
-            "org.jetbrains.kotlin.jvm"          : "${projectVersion.kotlin}",
-            "com.github.dcendents.android-maven": "2.1",
-            "com.gradle.plugin-publish"         : "0.13.0",
-            "org.jetbrains.dokka"               : "1.4.30",
-            "com.vanniktech.maven.publish"      : "0.14.2",
+            "org.jetbrains.kotlin.android": "${projectVersion.kotlin}",
+            "org.jetbrains.kotlin.jvm"    : "${projectVersion.kotlin}",
+            "com.gradle.plugin-publish"   : "0.15.0",
+            "org.jetbrains.dokka"         : "1.5.0",
+            "com.vanniktech.maven.publish": "0.14.2",
     ]
 }
